+++ src/main/java/com/chargebee/Environment.java
@@ -28,11 +28,11 @@ public class Environment {
      */
     public  int readTimeout = Integer.getInteger("com.chargebee.api.http.timeout.read", 60000);
 
-    public static String CHARSET = "UTF-8";
+    public static final String CHARSET = "UTF-8";
 
     public static final String API_VERSION = "v1";
     
-    public static final String LIBRARY_VERSION = "1.2.3";
+    public static final String LIBRARY_VERSION = "1.1.48";
 
     private final String apiBaseUrl;
 
@@ -53,7 +53,7 @@ public class Environment {
 
     public static Environment defaultConfig() {
         if(defaultEnv == null) {
-            throw new RuntimeException("Not yet configured");
+            throw new RuntimeException("The default environment has not been configured");
         }
         return defaultEnv;
     }