+++ src/main/java/com/chargebee/internal/Resource.java
@@ -1,8 +1,7 @@
 package com.chargebee.internal;
 
-import com.chargebee.Environment;
+import com.chargebee.gdata.PercentEscaper;
 import java.io.*;
-import java.net.URLEncoder;
 import java.sql.Timestamp;
 import java.util.*;
 import java.util.logging.Level;
@@ -206,6 +205,14 @@ public class Resource<T> {
             return null;
         }
         if(!type.isAssignableFrom(val.getClass())){
+            //JSON returns Integer values if the string format is without decimal points
+            // Like 10  instead of 10.0 
+            if(Double.class == type && val instanceof Number){
+                return (T) new Double(((Number)val).doubleValue());
+            }
+            if(Float.class == type && val instanceof Number){
+                return (T) new Float(((Number)val).floatValue());
+            }
             throw new RuntimeException("Type mismatch for property " + key
                     + " . Expected " + type.getName() + " but contains " + val.getClass().getName());
         }
@@ -250,13 +257,13 @@ public class Resource<T> {
         StringBuilder strBuf = new StringBuilder();
         for (String path : paths) {
             try {//Using URLEncoder is wrong as it encodes for form. Replace it with Google's CharEscapers.java
-                strBuf.append('/').append(URLEncoder.encode(path, Environment.CHARSET));
-            } catch (UnsupportedEncodingException ex) {
+                strBuf.append('/').append(new PercentEscaper(PercentEscaper.SAFEPATHCHARS_URLENCODER, false) .escape(path));
+            } catch (Exception ex) {
                 throw new RuntimeException(ex);
             }
         }
         return strBuf.toString();
-    }
-
+    }   
+    
 
 }