+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+### v1.1.48 (2014-09-16)
+* * *
+
+**Error Model**:
+
+New simpler model for error handling has been implemented. Please see https://apidocs.chargebee.com/docs/api?lang=java#error_handling for more details.
+The following variables in ApiException have been deprecated.
+* code  (Use apiErrorCode instead).
+* message (Use getMessage instead).
+* httpCode (Use httpStatusCode instead).
+
+**APIs Updated**:
+
+Shipping Address support added to *create subscription for a customer* api call.
+
 ### v1.1.47 (2014-08-28)
 * * *
 * Customer id can be passed to the checkout new subscription operation.